requirements.txt
yt-dlp>=2023.10.13
google-generativeai>=0.4.0
python-dotenv>=1.0.0
